+++ lib/src/api/event/player_view/picture_in_picture_enter_event.dart
@@ -4,7 +4,7 @@ import 'package:json_annotation/json_annotation.dart';
 
 part 'picture_in_picture_enter_event.g.dart';
 
-/// Is called when the [PlayerView] is about to enter Picture in Picture mode.
+/// Is called when the [PlayerView] is about to enter Picture-in-Picture mode.
 @JsonSerializable(explicitToJson: true)
 class PictureInPictureEnterEvent extends Event with EquatableMixin {
   const PictureInPictureEnterEvent({required super.timestamp});